requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==3.2.17
django-filter==22.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
Markdown==3.4.1
pycodestyle==2.10.0
pymongo==3.12.3
pytz==2022.7.1
sqlparse==0.2.4
tomli==2.0.1
tzdata==2022.7
